pyproject.toml
[project]
name = "fapigraphproject"
version = "1.0.0"
description = ""
authors = [
    {name = "Ya",email = "toomanycreepers@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.11,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "sqlalchemy (>=2.0.38,<3.0.0)",
    "sqlmodel (>=0.0.24,<0.0.25)",
    "alembic (>=1.15.1,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pyjwt[crypto] (>=2.10.1,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "strawberry-graphql (>=0.262.5,<0.263.0)",
    "websockets (>=15.0.1,<16.0.0)",
    "flake8 (>=7.2.0,<8.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "isort (>=6.0.1,<7.0.0)",
    "pre-commit (>=4.2.0,<5.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.production.dependencies]
gunicorn = "^23.0.0"

[tool.isort]
profile="black"

[tool.poetry.group.development.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
pytest = "^8.3.5"
